Move sudoku helper into sudoku component folder
